# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Client {
  id: string;
  fullName: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  leadSourceChannel: string;
  /** @nullable */
  leadSourceDetail?: string | null;
  createdAt: string;
}

export interface ClientInput {
  /** @minLength 1 */
  fullName: string;
  phone?: string;
  email?: string;
  /** @minLength 1 */
  leadSourceChannel: string;
  leadSourceDetail?: string;
}

export interface Case {
  id: string;
  caseNumber: number;
  clientId: string;
  clientName: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  ticketNumber?: string | null;
  /** @nullable */
  statuteCode?: string | null;
  /** @nullable */
  offenceDescription?: string | null;
  /** @nullable */
  officeCode?: string | null;
  /** @nullable */
  courtLocation?: string | null;
  intakeDate: string;
  offenceDate: string;
  status: string;
  totalFee: number;
  amountReceived: number;
  balanceOwing: number;
  /** @nullable */
  nextFollowUpDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CaseInput {
  clientId: string;
  intakeDate: string;
  offenceDate: string;
  ticketNumber?: string;
  statuteCode?: string;
  offenceDescription?: string;
  officeCode?: string;
  courtLocation?: string;
  status: string;
  /** @minimum 0 */
  totalFee: number;
  nextFollowUpDate?: string;
  firstNote?: string;
}

export interface CaseUpdate {
  intakeDate?: string;
  offenceDate?: string;
  ticketNumber?: string;
  statuteCode?: string;
  offenceDescription?: string;
  officeCode?: string;
  courtLocation?: string;
  status?: string;
  /** @minimum 0 */
  totalFee?: number;
  /** @nullable */
  nextFollowUpDate?: string | null;
}

export interface Payment {
  id: string;
  caseId: string;
  amount: number;
  date: string;
  /** @nullable */
  method?: string | null;
  /** @nullable */
  note?: string | null;
}

export interface PaymentInput {
  /** @minimum 0.01 */
  amount: number;
  date: string;
  method?: string;
  note?: string;
}

export interface Note {
  id: string;
  text: string;
  /** @nullable */
  author?: string | null;
  createdAt: string;
}

export interface NoteInput {
  /** @minLength 1 */
  text: string;
}

export interface CourtDate {
  id: string;
  date: string;
  /** @nullable */
  outcome?: string | null;
}

export type CaseDetail = Case & {
  notes: Note[];
  courtDates: CourtDate[];
  payments: Payment[];
};

export type DashboardStatusCounts = {[key: string]: number};

export interface Dashboard {
  activeCases: number;
  outstandingBalance: number;
  dueFollowUps: number;
  attentionCases: Case[];
  statusCounts: DashboardStatusCounts;
}

export interface ReportCount {
  label: string;
  count: number;
}

export interface ReportsSummary {
  leadSources: ReportCount[];
  referralBreakdown: ReportCount[];
  statusCounts: ReportCount[];
  outstandingCases: Case[];
}

export type ImportInputRowsItem = {[key: string]: string};

export type ImportInputMapping = {[key: string]: string};

export interface ImportInput {
  rows: ImportInputRowsItem[];
  mapping?: ImportInputMapping;
}

export type ImportPreviewSampleItem = {[key: string]: string};

export interface ImportPreview {
  rowCount: number;
  sample: ImportPreviewSampleItem[];
}

export interface ImportResult {
  imported: number;
  skipped: number;
}

export type ListClientsParams = {
search?: string;
};

export type ListCasesParams = {
search?: string;
status?: string;
source?: string;
outstanding?: boolean;
dateFrom?: string;
dateTo?: string;
};