# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Case {
  id: string;
  caseNumber: number;
  clientId: string;
  clientName: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  ticketNumber?: string | null;
  /** @nullable */
  statuteCode?: string | null;
  /** @nullable */
  offenceDescription?: string | null;
  /** @nullable */
  officeCode?: string | null;
  /** @nullable */
  courtLocation?: string | null;
  intakeDate: Date;
  offenceDate: Date;
  status: string;
  totalFee: number;
  amountReceived: number;
  balanceOwing: number;
  /** @nullable */
  nextFollowUpDate?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}