# lib/api-zod/src/generated/types/caseDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Case } from './case';
import type { CourtDate } from './courtDate';
import type { Note } from './note';
import type { Payment } from './payment';

export type CaseDetail = Case & {
  notes: Note[];
  courtDates: CourtDate[];
  payments: Payment[];
};