# lib/api-zod/src/generated/types/caseInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CaseInput {
  clientId: string;
  intakeDate: Date;
  offenceDate: Date;
  ticketNumber?: string;
  statuteCode?: string;
  offenceDescription?: string;
  officeCode?: string;
  courtLocation?: string;
  status: string;
  /** @minimum 0 */
  totalFee: number;
  nextFollowUpDate?: Date;
  firstNote?: string;
}