# lib/api-zod/src/generated/types/client.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Client {
  id: string;
  fullName: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  leadSourceChannel: string;
  /** @nullable */
  leadSourceDetail?: string | null;
  createdAt: Date;
}