# lib/api-zod/src/generated/types/clientInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClientInput {
  /** @minLength 1 */
  fullName: string;
  phone?: string;
  email?: string;
  /** @minLength 1 */
  leadSourceChannel: string;
  leadSourceDetail?: string;
}