# lib/api-zod/src/generated/types/courtDate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CourtDate {
  id: string;
  date: Date;
  /** @nullable */
  outcome?: string | null;
}