# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Case } from './case';
import type { DashboardStatusCounts } from './dashboardStatusCounts';

export interface Dashboard {
  activeCases: number;
  outstandingBalance: number;
  dueFollowUps: number;
  attentionCases: Case[];
  statusCounts: DashboardStatusCounts;
}