# lib/api-zod/src/generated/types/importInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ImportInputMapping } from './importInputMapping';
import type { ImportInputRowsItem } from './importInputRowsItem';

export interface ImportInput {
  rows: ImportInputRowsItem[];
  mapping?: ImportInputMapping;
}