# lib/api-zod/src/generated/types/importPreview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ImportPreviewSampleItem } from './importPreviewSampleItem';

export interface ImportPreview {
  rowCount: number;
  sample: ImportPreviewSampleItem[];
}