# lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Note {
  id: string;
  text: string;
  /** @nullable */
  author?: string | null;
  createdAt: Date;
}