# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Payment {
  id: string;
  caseId: string;
  amount: number;
  date: Date;
  /** @nullable */
  method?: string | null;
  /** @nullable */
  note?: string | null;
}