# lib/api-zod/src/generated/types/paymentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentInput {
  /** @minimum 0.01 */
  amount: number;
  date: Date;
  method?: string;
  note?: string;
}