# lib/api-zod/src/generated/types/reportsSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Case } from './case';
import type { ReportCount } from './reportCount';

export interface ReportsSummary {
  leadSources: ReportCount[];
  referralBreakdown: ReportCount[];
  statusCounts: ReportCount[];
  outstandingCases: Case[];
}